Updates as of 3/18/2020:
1) Updates the arm_node documentation to have more general wording
2) Added a feature that if a motor's operating mode was changed, the operating mode of any of its shadow motors will change as well (useful for joints controlled with two motors - like the 'elbow' or 'shoulder' joints on some robots
3) Added ability to torque on/off specified joints in the Python API
4) Instead of specifying a motor by 'id' for the RegisterValues service, you can now specify by its name (as defined in the motor config file)
5) Added ability to control an Arm using an Xbox360 controller per PR #15, though this feature is untested
6) Added 'arm_run' as a boolean launch file argument to many of the interbotix_examples ROS packages - this allows you to not include the arm_run.launch file (if for example, you'd like to run your own version of it) per Issue #10

Bug Fixes as of 3/18/2020:
1) Fixed bug where if the operating mode of a dual-motor joint was changed via the RegisterValues service, only one of the two motors would end up changing to the correct operating mode.

NOTES

A new version of this repository will soon be created that is more modular but will contain the same features - just with a slightly different naming convention. It will rely on the interbotix_ros_core repository which will contain the interbotix_sdk ROS package. This way, instead of maintaining the same ROS package (i.e. the interbotix_sdk package) in multiple repositories, it will only need to be maintained in one repository. Once the new repo is created, this repository will still be available, but will be considered deprecated.